他ユーザーの記事一覧取得ロジック変更

+++ src/main/java/co/jp/wever/graphql/application/datamodel/response/query/skill/SkillResponse.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class SkillResponse {
     private String id;
     private String name;
-    private int relatedCount;
-    private int level;
+    private long relatedCount;
+    private long level;
 }
